Rename type files
